Simplify ProcessHandle, now works only on Windows
